banking_backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
python-dateutil>=2.8.2
cryptography==42.0.5
django-auditlog==3.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
dj-database-url==2.1.0
django-prometheus==2.3.1
django-health-check==3.18.1
psutil==5.9.0
sentry-sdk==1.40.0
structlog==23.2.0
django-structlog==8.1.0
concurrent-log-handler==0.9.24
drf-spectacular==0.27.2
drf-spectacular[sidecar]==0.27.2
bleach==6.1.0
setuptools>=65.0.0
gunicorn==21.2.0
whitenoise==6.6.0